move nav tree data in models instead of services as it is not a service
