+++ src/app/core/models/left-nav-loading/item-nav-tree-types.ts
@@ -1,5 +1,5 @@
 import { NavMenuItem } from '../../http-services/item-navigation.service';
-import { NavTreeElement } from '../../services/left-nav-loading/nav-tree-data';
+import { NavTreeElement } from './nav-tree-data';
 
 export function isANavMenuItem(e: NavTreeElement): e is NavMenuItem {
   return 'attemptId' in e;
